+++ auth-service/src/main/java/org/sixpang/authservice/application/client/UserClient.java
@@ -1,0 +1,7 @@
+package org.sixpang.authservice.application.client;
+
+import org.sixpang.authservice.application.client.dto.UserAuthDto;
+
+public interface UserClient {
+    UserAuthDto getUserByEmail(String email);
+}
