+++ auth-service/src/main/java/org/sixpang/authservice/application/security/TokenService.java
@@ -1,0 +1,7 @@
+package org.sixpang.authservice.application.security;
+
+import java.util.UUID;
+
+public interface TokenService {
+    String createToken(UUID userId, String role);
+}
