+++ auth-service/src/main/java/org/sixpang/authservice/infrastructure/client/FeignUserClient.java
@@ -1,13 +1,12 @@
 package org.sixpang.authservice.infrastructure.client;
 
 import org.sixpang.commonserver.response.ApiResponse;
-import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.sixpang.authservice.infrastructure.client.dto.UserAuthDto;
+import org.sixpang.authservice.application.client.dto.UserAuthDto;
 
-@FeignClient(name = "user-service")
-public interface UserClient {
+@org.springframework.cloud.openfeign.FeignClient(name = "user-service")
+public interface FeignUserClient {
 
     @GetMapping("/api/users/email")
     ApiResponse<UserAuthDto> getUserByEmail(@RequestParam String email);
