+++ auth-service/src/main/java/org/sixpang/authservice/infrastructure/client/UserClientImpl.java
@@ -1,0 +1,18 @@
+package org.sixpang.authservice.infrastructure.client;
+
+import lombok.RequiredArgsConstructor;
+import org.sixpang.authservice.application.client.UserClient;
+import org.sixpang.authservice.application.client.dto.UserAuthDto;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class UserClientImpl implements UserClient {
+
+    private final FeignUserClient feignUserClient;
+
+    @Override
+    public UserAuthDto getUserByEmail(String email) {
+        return feignUserClient.getUserByEmail(email).getData();
+    }
+}
